Added clone capibilites to EngineComponent. Started working on dragging stuff. Also fixed the minor copy error so now power strip works correctly.

+++ se.liu.ida.l2.tddd78.psc/src/graphics/PSCGraphics.java
@@ -13,7 +13,7 @@ public interface PSCGraphics
 	 *
 	 * @return the virtual x-coordinate that corresponds to the specified screen x-coordinate
 	 */
-	public float getVirtualX(int screenX);
+	public int getVirtualX(int screenX);
 
 	/**
 	 * Returns the specified screen y-position converted to the corresponding virtual y-position
@@ -22,5 +22,5 @@ public interface PSCGraphics
 	 *
 	 * @return the virtual y-coordinate that corresponds to the specified screen y-coordinate
 	 */
-	public float getVirtualY(int screenY);
+	public int getVirtualY(int screenY);
 }
